requirements.txt
pyautogui>=0.9.54
opencv-python>=4.8.0
numpy>=1.24.0
PyQt5>=5.15.9
Pillow>=10.0.0
pygetwindow>=0.0.9
pywin32>=306
mss>=9.0.1
